+++ packages/components/src/components/TextArea/stories/Default.stories.tsx
@@ -69,6 +69,22 @@ export const ShowCharacterCount: Story = {
   args: { showCharacterCount: true, maxLength: 100 },
 };
 
-export const Resizeable: Story = {
+export const AutoResizeable: Story = {
   args: { rows: 1, autoResizeMaxRows: 5 },
+};
+
+export const Resizeable: Story = {
+  args: { allowResize: true },
+};
+
+export const HorizontallyResizeable: Story = {
+  args: { allowHorizontalResize: true },
+};
+
+export const VerticallyResizeable: Story = {
+  args: { allowVerticalResize: true },
+};
+
+export const VerticallyAndAutoResizeable: Story = {
+  args: { allowVerticalResize: true, rows: 1, autoResizeMaxRows: 5 },
 };